+++ changelog.md
@@ -1,12 +1,13 @@
 ## Todo
 - fix dim_info, abandon fslorient
 - SelectScan.vue to be async
-- save and load config
 - adapt to different screen size
 - after deselecting a study folder, any relevant scans should also be removed
 - automatic BISD conversion
     - 为了在扫描时可以直接出BIDS结果，可以每隔5秒检测一下有没有日期时间比“已选中的study中最老的”还要更新的study，全选其scan，增量转化（即只转化新增的，维护一个哈希表，如果不存在则转换，否则跳过）
 ## changelog
+- save, load and delete config files
+- separate the template of Vue Component into an HTML file
 - abandon dicomifier
 - get to post
 - select all and deselect all for scans